third_party/WebKit/LayoutTests/fast/viewport/viewport-112-expected.txt

CONSOLE ERROR: line 3: The key "." is not recognized and ignored.
ALERT: viewport size 400x440 scale 0.8 with limits [0.8, 5] and userScalable true
